Fixed the build on macos that was failing because of an ambigous call when passing an double to abs() that is expecting either an int or a long.

+++ src/backend/datasources/filters/SpiceReader.cpp
@@ -185,7 +185,7 @@ int SpiceFileReader::readData(std::vector<void*>& data, int skipLines, int maxLi
 				if (!mNgspice && mLTSpiceBug) {
 					// Bug in the ltspice binary raw format
 					// For more information see MR !108
-					value = abs(value);
+					value = abs((long)value);
 				}
 
 				(*static_cast<QVector<double>*>(data[0]))[patchesIndexOffset + l] = value;
